init diagram

+++ back/src/main/java/com/clevory/back/model/editor/Diagram.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 @Getter
 @Setter
 public class Diagram {
-    private int diagramId; // topology id
+    private long diagramId; // topology id
     private String name; // topology name
     private ArrayList<Node> nodes;
     private ArrayList<Link> links;
